Logging added

+++ CodeAcademy.CoreWebApi/Controllers/Administrator/EditorsController.cs
@@ -7,6 +7,7 @@ using CodeAcademy.CoreWebApi.DataAccessLayer.AppIdentity;
 using CodeAcademy.CoreWebApi.DataTransferObject;
 using CodeAcademy.CoreWebApi.Helpers;
 using CodeAcademy.CoreWebApi.Helpers.Extensions;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.DependencyInjection;
@@ -14,26 +15,33 @@ using Microsoft.Extensions.Options;
 
 namespace CodeAcademy.CoreWebApi.Controllers.Administrator
 {
+    [Authorize]
     [Route("api/[controller]")]
     [ApiController]
     public class EditorsController : ControllerBase
     {
         private IAuthRepository _context;
+        private IAppRepository _app;
         private IOptions<CloudinarySettings> _cloudinaryConfig;
 
-        public EditorsController(IAuthRepository context,
+        public EditorsController(IAuthRepository context,IAppRepository app,
                                       IOptions<CloudinarySettings> cloudinaryConfig)
         {
             this._cloudinaryConfig = cloudinaryConfig;
             this._context = context;
+            this._app = app;
         }
 
         [HttpGet]
         [Route("getall")]
         public async Task<IActionResult> GetAll()
         {
-            var result = await _context.GetUsersByRole("Editor");
-            return Ok(result);
+            if (this.ValidRoleForAction(_app, _context, new string[] { "Admin" }))
+            {
+                var result = await _context.GetUsersByRole("Editor");
+                return Ok(result);
+            }
+            return Forbid();
         }
 
 
@@ -42,37 +50,41 @@ namespace CodeAcademy.CoreWebApi.Controllers.Administrator
         [Route("add")]
         public async Task<IActionResult> Add([FromForm] EditorModel model)
         {
-            bool saved;
-            if (ModelState.IsValid)
+            if (this.ValidRoleForAction(_app, _context, new string[] { "Admin" }))
             {
-                AppIdentityUser editor = new AppIdentityUser
+                bool saved;
+                if (ModelState.IsValid)
                 {
-                    Name = model.Name,
-                    Surname = model.Surname,
-                    BirthDate = model.BirthDate,
-                    Email = model.Email,
-                    GenderId = model.GenderId,
-                    PhotoId = 1,
-                    UserName = model.Email
-                };
-                if (await _context.FindUserByEmail(model.Email) == null)
-                {
-                    await _context.Register(editor, model.Password);
-                    saved = await _context.AddUserToRole(editor, "Editor") != null;
-                    if (saved == true)
+                    AppIdentityUser editor = new AppIdentityUser
                     {
-                        var urlHelper = HttpContext.RequestServices.GetRequiredService<IUrlHelper>();
-                        await this.SendConfirmaitionMail(editor, _context, urlHelper);
-                        return Ok(editor);
+                        Name = model.Name,
+                        Surname = model.Surname,
+                        BirthDate = model.BirthDate,
+                        Email = model.Email,
+                        GenderId = model.GenderId,
+                        PhotoId = 1,
+                        UserName = model.Email
+                    };
+                    if (await _context.FindUserByEmail(model.Email) == null)
+                    {
+                        await _context.Register(editor, model.Password);
+                        saved = await _context.AddUserToRole(editor, "Editor") != null;
+                        if (saved == true)
+                        {
+                            var urlHelper = HttpContext.RequestServices.GetRequiredService<IUrlHelper>();
+                            await this.SendConfirmaitionMail(editor, _context, urlHelper);
+                            return Ok(editor);
+                        }
                     }
-                }
-                else
-                {
-                    return BadRequest($"User with {model.Email} email already exists");
-                }
+                    else
+                    {
+                        return BadRequest($"User with {model.Email} email already exists");
+                    }
 
+                }
+                return BadRequest("Model is not valid");
             }
-            return BadRequest("Model is not valid");
+            return Forbid();
         }
 
 
@@ -80,21 +92,24 @@ namespace CodeAcademy.CoreWebApi.Controllers.Administrator
         [Route("delete")]
         public async Task<IActionResult> Delete([FromForm] string id)
         {
-
-            AppIdentityUser item = await _context.FindUserById(id);
-            if (item != null)
+            if (this.ValidRoleForAction(_app, _context, new string[] { "Admin" }))
             {
-                bool result = await _context.DeleteUser(item) != null;
+                AppIdentityUser item = await _context.FindUserById(id);
+                if (item != null)
+                {
+                    bool result = await _context.DeleteUser(item) != null;
 
-                if (result == true)
-                    return Ok(item);
+                    if (result == true)
+                        return Ok(item);
+                    else
+                        return BadRequest("Model cannot be  deleted");
+                }
                 else
-                    return BadRequest("Model cannot be  deleted");
+                {
+                    return NotFound("Model not found");
+                }
             }
-            else
-            {
-                return NotFound("Model not found");
-            }
+            return Forbid();
         }
 
 
@@ -102,27 +117,30 @@ namespace CodeAcademy.CoreWebApi.Controllers.Administrator
         [Route("update")]
         public async Task<IActionResult> Update([FromForm] EditorModel model)
         {
-            bool saved;
-            if (ModelState.IsValid)
+            if (this.ValidRoleForAction(_app, _context, new string[] { "Admin" }))
             {
-                AppIdentityUser item = await _context.FindUserById(model.Id);
-
-                item.Name = model.Name;
-                item.Surname = model.Surname;
-                item.BirthDate = model.BirthDate;
-
-                saved = await _context.UpdateUser(item) != null;
-                if (saved == true)
+                bool saved;
+                if (ModelState.IsValid)
                 {
-                    return Ok(item);
+                    AppIdentityUser item = await _context.FindUserById(model.Id);
+
+                    item.Name = model.Name;
+                    item.Surname = model.Surname;
+                    item.BirthDate = model.BirthDate;
+
+                    saved = await _context.UpdateUser(item) != null;
+                    if (saved == true)
+                    {
+                        return Ok(item);
+                    }
+                    else
+                    {
+                        return BadRequest("Item cannot be updated");
+                    }
                 }
-                else
-                {
-                    return BadRequest("Item cannot be updated");
-                }
+                return BadRequest("Model is not valid");
             }
-            return BadRequest("Model is not valid");
-
+            return Forbid();
         }
     }
 }
