+++ CodeAcademy.CoreWebApi/Controllers/Administrator/FacultiesController.cs
@@ -7,32 +7,41 @@ using CodeAcademy.CoreWebApi.DataAccessLayer.Entities;
 using CodeAcademy.CoreWebApi.DataTransferObject;
 using CodeAcademy.CoreWebApi.Entities;
 using CodeAcademy.CoreWebApi.Helpers;
+using CodeAcademy.CoreWebApi.Helpers.Extensions;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Options;
 
 namespace CodeAcademy.CoreWebApi.Controllers.Administrator
 {
+    [Authorize]
     [Route("api/[controller]")]
     [ApiController]
     public class FacultiesController : ControllerBase
     {
         private IAppRepository _context;
+        private IAuthRepository _auth;
         private IOptions<CloudinarySettings> _cloudinaryConfig;
 
-        public FacultiesController(IAppRepository context,
+        public FacultiesController(IAppRepository context, IAuthRepository auth,
                                       IOptions<CloudinarySettings> cloudinaryConfig)
         {
             this._cloudinaryConfig = cloudinaryConfig;
             this._context = context;
+            this._auth = auth;
         }
 
         [HttpGet]
         [Route("getall")]
         public async Task<IActionResult> GetAll()
         {
-            var result = await _context.GetFacultiesAsync();
-            return Ok(result);
+            if (this.ValidRoleForAction(_context, _auth, new string[] { "Admin" }))
+            {
+                var result = await _context.GetFacultiesAsync();
+                return Ok(result);
+            }
+            return Forbid();
         }
 
 
@@ -40,28 +49,32 @@ namespace CodeAcademy.CoreWebApi.Controllers.Administrator
         [Route("add")]
         public async Task<IActionResult> Add([FromForm] FacultyModel model)
         {
-            bool saved;
-            if (ModelState.IsValid)
+            if (this.ValidRoleForAction(_context, _auth, new string[] { "Admin" }))
             {
-                PhotoUploadCloudinary upload = new PhotoUploadCloudinary(_cloudinaryConfig);
-                Photo photo = upload.Upload(model.Photo);
-
-                Faculty item = new Faculty
+                bool saved;
+                if (ModelState.IsValid)
                 {
-                    Name = model.Name,
-                    LessonHour = model.LessonHour,
-                    Photo = photo
-                };
+                    PhotoUploadCloudinary upload = new PhotoUploadCloudinary(_cloudinaryConfig);
+                    Photo photo = upload.Upload(model.Photo);
 
-                await _context.Add(photo);
-                await _context.Add(item);
-                saved = _context.SaveAll();
-                if (saved == true)
-                {
-                    return Ok(item);
+                    Faculty item = new Faculty
+                    {
+                        Name = model.Name,
+                        LessonHour = model.LessonHour,
+                        Photo = photo
+                    };
+
+                    await _context.Add(photo);
+                    await _context.Add(item);
+                    saved = await _context.SaveAll();
+                    if (saved == true)
+                    {
+                        return Ok(item);
+                    }
                 }
+                return BadRequest("Model is not valid");
             }
-            return BadRequest("Model is not valid");
+            return Forbid();
         }
 
 
@@ -69,22 +82,25 @@ namespace CodeAcademy.CoreWebApi.Controllers.Administrator
         [Route("delete")]
         public async Task<IActionResult> Delete([FromForm] int id)
         {
-
-            Faculty item = await _context.GetByIdAsync<Faculty>(x => x.Id == id);
-            if (item != null)
+            if (this.ValidRoleForAction(_context, _auth, new string[] { "Admin" }))
             {
-                _context.Delete(item);
-                bool result = _context.SaveAll();
+                Faculty item = await _context.GetByIdAsync<Faculty>(x => x.Id == id);
+                if (item != null)
+                {
+                    _context.Delete(item);
+                    bool result = await _context.SaveAll();
 
-                if (result == true)
-                    return Ok(item);
+                    if (result == true)
+                        return Ok(item);
+                    else
+                        return BadRequest("Model cannot be  deleted");
+                }
                 else
-                    return BadRequest("Model cannot be  deleted");
+                {
+                    return NotFound("Model not found");
+                }
             }
-            else
-            {
-                return NotFound("Model not found");
-            }
+            return Forbid();
         }
 
 
@@ -92,31 +108,34 @@ namespace CodeAcademy.CoreWebApi.Controllers.Administrator
         [Route("update")]
         public async Task<IActionResult> Update([FromForm] FacultyModel model)
         {
-            bool saved;
-            if (ModelState.IsValid)
+            if (this.ValidRoleForAction(_context, _auth, new string[] { "Admin" }))
             {
-                Faculty item = await _context.GetByIdAsync<Faculty>(x => x.Id == model.Id);
-                PhotoUploadCloudinary upload = new PhotoUploadCloudinary(_cloudinaryConfig);
-                Photo photo = upload.Upload(model.Photo);
-
-                item.Name = model.Name;
-                item.LessonHour = model.LessonHour;
-                item.Photo = photo;
-
-                await _context.Add(photo);
-                _context.Update(item);
-                saved = _context.SaveAll();
-                if (saved == true)
+                bool saved;
+                if (ModelState.IsValid)
                 {
-                    return Ok(item);
+                    Faculty item = await _context.GetByIdAsync<Faculty>(x => x.Id == model.Id);
+                    PhotoUploadCloudinary upload = new PhotoUploadCloudinary(_cloudinaryConfig);
+                    Photo photo = upload.Upload(model.Photo);
+
+                    item.Name = model.Name;
+                    item.LessonHour = model.LessonHour;
+                    item.Photo = photo;
+
+                    await _context.Add(photo);
+                    _context.Update(item);
+                    saved = await _context.SaveAll();
+                    if (saved == true)
+                    {
+                        return Ok(item);
+                    }
+                    else
+                    {
+                        return BadRequest("Item cannot be updated");
+                    }
                 }
-                else
-                {
-                    return BadRequest("Item cannot be updated");
-                }
+                return BadRequest("Model is not valid");
             }
-            return BadRequest("Model is not valid");
-
+            return Forbid();
         }
     }
 }
