+++ CodeAcademy.CoreWebApi/Controllers/Administrator/LeftNavItemsController.cs
@@ -4,61 +4,74 @@ using CodeAcademy.CoreWebApi.DataAccessLayer.Entities;
 using CodeAcademy.CoreWebApi.DataTransferObject;
 using CodeAcademy.CoreWebApi.Entities;
 using CodeAcademy.CoreWebApi.Helpers;
+using CodeAcademy.CoreWebApi.Helpers.Extensions;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Options;
 
 namespace CodeAcademy.CoreWebApi.Controllers.Administrator
 {
+    [Authorize]
     [Route("api/[controller]")]
     [ApiController]
     public class LeftNavItemsController : ControllerBase
     {
         private IAppRepository _context;
+        private IAuthRepository _auth;
         private IOptions<CloudinarySettings> _cloudinaryConfig;
 
-        public LeftNavItemsController(IAppRepository context,
+        public LeftNavItemsController(IAppRepository context, IAuthRepository auth,
                                       IOptions<CloudinarySettings> cloudinaryConfig)
         {
             this._cloudinaryConfig = cloudinaryConfig;
             this._context = context;
+            this._auth = auth;
         }
 
         [HttpGet]
         [Route("getall")]
         public async Task<IActionResult> GetAll()
         {
-            var result = await _context.GetLeftNavItems();
-            return Ok(result);
+            if (this.ValidRoleForAction(_context, _auth, new string[] { "Admin" }))
+            {
+                var result = await _context.GetLeftNavItems();
+                return Ok(result);
+            }
+            return Forbid();
         }
 
         [HttpPost]
         [Route("add")]
         public async Task<IActionResult> Add([FromForm] LeftNavItemModel model)
         {
-            bool saved;
-            if (ModelState.IsValid)
+            if (this.ValidRoleForAction(_context, _auth, new string[] { "Admin" }))
             {
-                PhotoUploadCloudinary upload = new PhotoUploadCloudinary(_cloudinaryConfig);
-                Photo photo = upload.Upload(model.Photo);
-
-                LeftNavItem item = new LeftNavItem
+                bool saved;
+                if (ModelState.IsValid)
                 {
-                    Name = model.Name,
-                    RouterLink = model.RouterLink,
-                    IconClassname = model.IconClassname,
-                    IsVisible = model.IsVisible,
-                    Photo = photo
-                };
+                    PhotoUploadCloudinary upload = new PhotoUploadCloudinary(_cloudinaryConfig);
+                    Photo photo = upload.Upload(model.Photo);
 
-                await _context.Add(photo);
-                await _context.Add(item);
-                saved = _context.SaveAll();
-                if(saved == true)
-                {
-                    return Ok(item);
+                    LeftNavItem item = new LeftNavItem
+                    {
+                        Name = model.Name,
+                        RouterLink = model.RouterLink,
+                        IconClassname = model.IconClassname,
+                        IsVisible = model.IsVisible,
+                        Photo = photo
+                    };
+
+                    await _context.Add(photo);
+                    await _context.Add(item);
+                    saved = await _context.SaveAll();
+                    if (saved == true)
+                    {
+                        return Ok(item);
+                    }
                 }
+                return BadRequest("Model is not valid");
             }
-            return BadRequest("Model is not valid");
+            return Forbid();
         }
 
 
@@ -66,22 +79,25 @@ namespace CodeAcademy.CoreWebApi.Controllers.Administrator
         [Route("delete")]
         public async Task<IActionResult> Delete([FromForm] int id)
         {
-
-            LeftNavItem item =  await _context.GetByIdAsync<LeftNavItem>(x => x.Id == id);
-            if(item != null)
+            if (this.ValidRoleForAction(_context, _auth, new string[] { "Admin" }))
             {
-                _context.Delete(item);
-                bool result = _context.SaveAll();
+                LeftNavItem item = await _context.GetByIdAsync<LeftNavItem>(x => x.Id == id);
+                if (item != null)
+                {
+                    _context.Delete(item);
+                    bool result = await _context.SaveAll();
 
-                if (result == true)
-                    return Ok(item);
+                    if (result == true)
+                        return Ok(item);
+                    else
+                        return BadRequest("Model cannot be  deleted");
+                }
                 else
-                    return BadRequest("Model cannot be  deleted");
+                {
+                    return NotFound("Model not found");
+                }
             }
-            else
-            {
-                return NotFound("Model not found");
-            }
+            return Forbid();
         }
 
 
@@ -89,33 +105,36 @@ namespace CodeAcademy.CoreWebApi.Controllers.Administrator
         [Route("update")]
         public async Task<IActionResult> Update([FromForm] LeftNavItemModel model)
         {
-            bool saved;
-            if (ModelState.IsValid)
+            if (this.ValidRoleForAction(_context, _auth, new string[] { "Admin" }))
             {
-                LeftNavItem item = await _context.GetByIdAsync<LeftNavItem>(x => x.Id == model.Id);
-                PhotoUploadCloudinary upload = new PhotoUploadCloudinary(_cloudinaryConfig);
-                Photo photo = upload.Upload(model.Photo);
-
-                item.Name = model.Name;
-                item.RouterLink = model.RouterLink;
-                item.IconClassname = model.IconClassname;
-                item.IsVisible = model.IsVisible;
-                item.Photo = photo;
-
-                await _context.Add(photo);
-                _context.Update(item);
-                saved = _context.SaveAll();
-                if (saved == true)
+                bool saved;
+                if (ModelState.IsValid)
                 {
-                    return Ok(item);
+                    LeftNavItem item = await _context.GetByIdAsync<LeftNavItem>(x => x.Id == model.Id);
+                    PhotoUploadCloudinary upload = new PhotoUploadCloudinary(_cloudinaryConfig);
+                    Photo photo = upload.Upload(model.Photo);
+
+                    item.Name = model.Name;
+                    item.RouterLink = model.RouterLink;
+                    item.IconClassname = model.IconClassname;
+                    item.IsVisible = model.IsVisible;
+                    item.Photo = photo;
+
+                    await _context.Add(photo);
+                    _context.Update(item);
+                    saved = await _context.SaveAll();
+                    if (saved == true)
+                    {
+                        return Ok(item);
+                    }
+                    else
+                    {
+                        return BadRequest("Item cannot be updated");
+                    }
                 }
-                else
-                {
-                    return BadRequest("Item cannot be updated");
-                }
+                return BadRequest("Model is not valid");
             }
-            return BadRequest("Model is not valid");
-
+            return Forbid();
         }
     }
 }
