+++ CodeAcademy.CoreWebApi/Controllers/Administrator/LessonHoursController.cs
@@ -6,17 +6,21 @@ using CodeAcademy.CoreWebApi.BusinessLogicLayer.Abstract;
 using CodeAcademy.CoreWebApi.DataAccessLayer.Entities;
 using CodeAcademy.CoreWebApi.DataTransferObject;
 using CodeAcademy.CoreWebApi.Helpers;
+using CodeAcademy.CoreWebApi.Helpers.Extensions;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Options;
 
 namespace CodeAcademy.CoreWebApi.Controllers.Administrator
 {
+    [Authorize]
     [Route("api/[controller]")]
     [ApiController]
     public class LessonHoursController : ControllerBase
     {
         private IAppRepository _context;
+        private IAuthRepository _auth;
         private IOptions<CloudinarySettings> _cloudinaryConfig;
 
         public LessonHoursController(IAppRepository context,
@@ -30,8 +34,12 @@ namespace CodeAcademy.CoreWebApi.Controllers.Administrator
         [Route("getall")]
         public async Task<IActionResult> GetAll()
         {
-            var result = await _context.GetLessonHoursAsync();
-            return Ok(result);
+            if (this.ValidRoleForAction(_context, _auth, new string[] { "Admin" }))
+            {
+                var result = await _context.GetLessonHoursAsync();
+                return Ok(result); 
+            }
+            return Forbid();
         }
 
 
@@ -40,33 +48,37 @@ namespace CodeAcademy.CoreWebApi.Controllers.Administrator
         [Route("add")]
         public async Task<IActionResult> Add([FromForm] LessonHourModel model)
         {
-            bool saved;
-            if (ModelState.IsValid)
+            if (this.ValidRoleForAction(_context, _auth, new string[] { "Admin" }))
             {
-                LessonHour item = new LessonHour
+                bool saved;
+                if (ModelState.IsValid)
                 {
-                    Name = model.Name,
-                    BeginHour = model.BeginHour,
-                    BeginMinute = model.BeginMinute,
-                    EndHour = model.EndHour,
-                    EndMinute = model.EndMinute,
-                    Friday = model.Friday,
-                    Monday = model.Monday,
-                    Saturday = model.Saturday,
-                    Sunday = model.Sunday,
-                    Thursday = model.Thursday,
-                    Tuesday = model.Tuesday,
-                    Wednesday = model.Wednesday
-                };
+                    LessonHour item = new LessonHour
+                    {
+                        Name = model.Name,
+                        BeginHour = model.BeginHour,
+                        BeginMinute = model.BeginMinute,
+                        EndHour = model.EndHour,
+                        EndMinute = model.EndMinute,
+                        Friday = model.Friday,
+                        Monday = model.Monday,
+                        Saturday = model.Saturday,
+                        Sunday = model.Sunday,
+                        Thursday = model.Thursday,
+                        Tuesday = model.Tuesday,
+                        Wednesday = model.Wednesday
+                    };
 
-                await _context.Add(item);
-                saved = _context.SaveAll();
-                if (saved == true)
-                {
-                    return Ok(item);
+                    await _context.Add(item);
+                    saved = await _context.SaveAll();
+                    if (saved == true)
+                    {
+                        return Ok(item);
+                    }
                 }
+                return BadRequest("Model is not valid");
             }
-            return BadRequest("Model is not valid");
+            return Forbid();
         }
 
 
@@ -74,21 +86,25 @@ namespace CodeAcademy.CoreWebApi.Controllers.Administrator
         [Route("delete")]
         public async Task<IActionResult> Delete([FromForm] int id)
         {
-            LessonHour item = await _context.GetByIdAsync<LessonHour>(x => x.Id == id);
-            if (item != null)
+            if (this.ValidRoleForAction(_context, _auth, new string[] { "Admin" }))
             {
-                _context.Delete(item);
-                bool result = _context.SaveAll();
+                LessonHour item = await _context.GetByIdAsync<LessonHour>(x => x.Id == id);
+                if (item != null)
+                {
+                    _context.Delete(item);
+                    bool result = await _context.SaveAll();
 
-                if (result == true)
-                    return Ok(item);
+                    if (result == true)
+                        return Ok(item);
+                    else
+                        return BadRequest("Model cannot be  deleted");
+                }
                 else
-                    return BadRequest("Model cannot be  deleted");
+                {
+                    return NotFound("Model not found");
+                }
             }
-            else
-            {
-                return NotFound("Model not found");
-            }
+            return Forbid();
         }
 
 
@@ -96,37 +112,40 @@ namespace CodeAcademy.CoreWebApi.Controllers.Administrator
         [Route("update")]
         public async Task<IActionResult> Update([FromForm] LessonHourModel model)
         {
-            bool saved;
-            if (ModelState.IsValid)
+            if (this.ValidRoleForAction(_context, _auth, new string[] { "Admin" }))
             {
-                LessonHour item = await _context.GetByIdAsync<LessonHour>(x => x.Id == model.Id);
+                bool saved;
+                if (ModelState.IsValid)
+                {
+                    LessonHour item = await _context.GetByIdAsync<LessonHour>(x => x.Id == model.Id);
 
-                item.Name = model.Name;
-                item.Wednesday = model.Wednesday;
-                item.Monday = model.Monday;
-                item.Saturday = model.Saturday;
-                item.Sunday = model.Sunday;
-                item.Thursday = model.Thursday;
-                item.Tuesday = model.Tuesday;
-                item.Friday = model.Friday;
-                item.BeginMinute = model.BeginMinute;
-                item.BeginHour = model.BeginHour;
-                item.EndHour = model.EndHour;
-                item.EndMinute = model.EndMinute;
-                
-                _context.Update(item);
-                saved = _context.SaveAll();
-                if (saved == true)
-                {
-                    return Ok(item);
+                    item.Name = model.Name;
+                    item.Wednesday = model.Wednesday;
+                    item.Monday = model.Monday;
+                    item.Saturday = model.Saturday;
+                    item.Sunday = model.Sunday;
+                    item.Thursday = model.Thursday;
+                    item.Tuesday = model.Tuesday;
+                    item.Friday = model.Friday;
+                    item.BeginMinute = model.BeginMinute;
+                    item.BeginHour = model.BeginHour;
+                    item.EndHour = model.EndHour;
+                    item.EndMinute = model.EndMinute;
+
+                    _context.Update(item);
+                    saved = await _context.SaveAll();
+                    if (saved == true)
+                    {
+                        return Ok(item);
+                    }
+                    else
+                    {
+                        return BadRequest("Item cannot be updated");
+                    }
                 }
-                else
-                {
-                    return BadRequest("Item cannot be updated");
-                }
+                return BadRequest("Model is not valid");
             }
-            return BadRequest("Model is not valid");
-
+            return Forbid();
         }
     }
 }
