+++ CodeAcademy.CoreWebApi/Controllers/Editor/TagsController.cs
@@ -7,29 +7,37 @@ using CodeAcademy.CoreWebApi.DataAccessLayer.Entities;
 using CodeAcademy.CoreWebApi.DataTransferObject;
 using CodeAcademy.CoreWebApi.Entities;
 using CodeAcademy.CoreWebApi.Helpers;
+using CodeAcademy.CoreWebApi.Helpers.Extensions;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Options;
 
 namespace CodeAcademy.CoreWebApi.Controllers.Editor
 {
+    [Authorize]
     [Route("api/[controller]")]
     [ApiController]
     public class TagsController : ControllerBase
     {
         private IAppRepository _context;
-
-        public TagsController(IAppRepository context)
+        private IAuthRepository _auth;
+        public TagsController(IAppRepository context, IAuthRepository auth)
         {
             this._context = context;
+            this._auth = auth;
         }
 
         [HttpGet]
         [Route("getall")]
         public async Task<IActionResult> GetAll()
         {
-            var result = await _context.GetTagsAsync();
-            return Ok(result);
+            if (this.ValidRoleForAction(_context, _auth, new string[] { "Editor" }))
+            {
+                var result = await _context.GetTagsAsync();
+                return Ok(result);
+            }
+            return Forbid();
         }
 
 
@@ -38,29 +46,33 @@ namespace CodeAcademy.CoreWebApi.Controllers.Editor
         [Route("add")]
         public async Task<IActionResult> Add([FromForm] TagModel model)
         {
-            bool saved;
-            if (ModelState.IsValid)
+            if (this.ValidRoleForAction(_context, _auth, new string[] { "Editor" }))
             {
-                if (await IsUniqueTag(model.Name,model.FacultyId))
+                bool saved;
+                if (ModelState.IsValid)
                 {
-                    Tag item = new Tag
+                    if (await IsUniqueTag(model.Name, model.FacultyId))
                     {
-                        Name = model.Name,
-                        FacultyId = model.FacultyId
-                    };
-                    await _context.Add(item);
-                    saved = _context.SaveAll();
-                    if (saved == true)
+                        Tag item = new Tag
+                        {
+                            Name = model.Name,
+                            FacultyId = model.FacultyId
+                        };
+                        await _context.Add(item);
+                        saved = await _context.SaveAll();
+                        if (saved == true)
+                        {
+                            return Ok(item);
+                        }
+                    }
+                    else
                     {
-                        return Ok(item);
+                        return BadRequest("This tag already exists in database");
                     }
                 }
-                else
-                {
-                    return BadRequest("This tag already exists in database");
-                }
+                return BadRequest("Model is not valid");
             }
-            return BadRequest("Model is not valid");
+            return Forbid();
         }
 
 
@@ -68,22 +80,25 @@ namespace CodeAcademy.CoreWebApi.Controllers.Editor
         [Route("delete")]
         public async Task<IActionResult> Delete([FromForm] int id)
         {
-
-            Tag item = await _context.GetByIdAsync<Tag>(x => x.Id == id);
-            if (item != null)
+            if (this.ValidRoleForAction(_context, _auth, new string[] { "Editor" }))
             {
-                _context.Delete(item);
-                bool result = _context.SaveAll();
+                Tag item = await _context.GetByIdAsync<Tag>(x => x.Id == id);
+                if (item != null)
+                {
+                    _context.Delete(item);
+                    bool result = await _context.SaveAll();
 
-                if (result == true)
-                    return Ok(item);
+                    if (result == true)
+                        return Ok(item);
+                    else
+                        return BadRequest("Model cannot be  deleted");
+                }
                 else
-                    return BadRequest("Model cannot be  deleted");
+                {
+                    return NotFound("Model not found");
+                }
             }
-            else
-            {
-                return NotFound("Model not found");
-            }
+            return Forbid();
         }
 
 
@@ -91,27 +106,30 @@ namespace CodeAcademy.CoreWebApi.Controllers.Editor
         [Route("update")]
         public async Task<IActionResult> Update([FromForm] TagModel model)
         {
-            bool saved;
-            if (ModelState.IsValid)
+            if (this.ValidRoleForAction(_context, _auth, new string[] { "Editor" }))
             {
-                Tag item = await _context.GetByIdAsync<Tag>(x => x.Id == model.Id);
-
-                item.Name = model.Name;
-                item.FacultyId = model.FacultyId;
-
-                _context.Update(item);
-                saved = _context.SaveAll();
-                if (saved == true)
+                bool saved;
+                if (ModelState.IsValid)
                 {
-                    return Ok(item);
+                    Tag item = await _context.GetByIdAsync<Tag>(x => x.Id == model.Id);
+
+                    item.Name = model.Name;
+                    item.FacultyId = model.FacultyId;
+
+                    _context.Update(item);
+                    saved = await _context.SaveAll();
+                    if (saved == true)
+                    {
+                        return Ok(item);
+                    }
+                    else
+                    {
+                        return BadRequest("Item cannot be updated");
+                    }
                 }
-                else
-                {
-                    return BadRequest("Item cannot be updated");
-                }
+                return BadRequest("Model is not valid");
             }
-            return BadRequest("Model is not valid");
-
+            return Forbid();
         }
 
         private async Task<bool> IsUniqueTag(string name, int facultyId)
